requirements.txt
sounddevice
vosk
numpy
pvporcupine
flask
flask-cors
dotenv
google-genai
google-api-core
Pillow
opencv-python